+++ client/Command.ts
@@ -16,6 +16,17 @@ module ImprovedInitiative {
         }
 
         this.ToolTip = ko.computed(() => `${this.Description} [${this.KeyBinding}]`);
+
+        let keyBinding = Store.Load<string>(Store.KeyBindings, this.Description);
+        if (keyBinding) {
+            this.KeyBinding = keyBinding;
+        }
+
+        let showOnActionBarSetting = Store.Load<boolean>(Store.ActionBar, this.Description);
+        if (showOnActionBarSetting != null) {
+            this.ShowOnActionBar(showOnActionBar);
+        }
+        
     }
   }
   
@@ -25,15 +36,13 @@ module ImprovedInitiative {
         new Command('Clear Encounter', c.ClearEncounter, 'alt+del', 'fa-trash'),
         new Command('Open Library', c.ShowLibraries, 'alt+a', 'fa-user-plus'),
         new Command('Show Player Window', c.LaunchPlayerWindow, 'alt+w', 'fa-users'),
-        new Command('Select Next Combatant', c.SelectNextCombatant, 'j', 'fa-arrow-down', false),
-        new Command('Select Previous Combatant', c.SelectPreviousCombatant, 'k', 'fa-arrow-up', false),
         new Command('Next Turn', c.NextTurn, 'n', 'fa-step-forward'),
         new Command('Previous Turn', c.PreviousTurn, 'alt+n', 'fa-step-backward'),
         new Command('Save Encounter', c.SaveEncounter, 'alt+s', 'fa-save'),
         new Command('Settings', c.ShowSettings, '?', 'fa-gear', true, true)
   ]
     
-    export var BuildCombatantCommandList: (c: EncounterCommander) => Command [] = c => [
+    export var BuildCombatantCommandList: (c: CombatantCommander) => Command [] = c => [
         new Command('Damage/Heal Selected Combatant', c.FocusSelectedCreatureHP, 't', 'fa-plus-circle'),
         new Command('Add Note to Selected Combatant', c.AddSelectedCreatureTag, 'g', 'fa-tag', false),
         new Command('Remove Selected Combatant from Encounter', c.RemoveSelectedCreatures, 'del', 'fa-remove'),
@@ -42,6 +51,10 @@ module ImprovedInitiative {
         new Command('Edit Selected Combatant Initiative', c.EditSelectedCreatureInitiative, 'alt+i', 'fa-play-circle-o'),
         new Command('Move Selected Combatant Down', c.MoveSelectedCreatureDown, 'alt+j', 'fa-angle-double-down'),
         new Command('Move Selected Combatant Up', c.MoveSelectedCreatureUp, 'alt+k', 'fa-angle-double-up'),
-        new Command('Apply Temporary HP', c.AddSelectedCreaturesTemporaryHP, 'alt+t', 'fa-medkit')
-    ]
+        new Command('Apply Temporary HP', c.AddSelectedCreaturesTemporaryHP, 'alt+t', 'fa-medkit'),
+        new Command('Select Next Combatant', c.SelectNextCombatant, 'j', 'fa-arrow-down', false),
+        new Command('Select Previous Combatant', c.SelectPreviousCombatant, 'k', 'fa-arrow-up', false)        
+  ]
+    
+    
 }
