+++ client/Settings.ts
@@ -1,6 +1,7 @@
 module ImprovedInitiative {
     interface Params {
         encounterCommander: EncounterCommander;
+        combatantCommander: CombatantCommander;
     }
     export var Settings = (params: Params) => {
         var tips = [
@@ -49,7 +50,9 @@ module ImprovedInitiative {
         displayTurnTimer.subscribe(params.encounterCommander.DisplayTurnTimer);
         
         return {
-            Commander: params.encounterCommander,
+            EncounterCommands: params.encounterCommander.Commands,
+            CombatantCommands: params.combatantCommander.Commands,
+            
             CurrentTab: ko.observable<string>('about'),
             ExportData: () => {
                 var blob = Store.ExportAll();
@@ -85,7 +88,8 @@ module ImprovedInitiative {
             PlayerViewDisplayTurnTimer: loadSetting("PlayerViewDisplayTurnTimer", false),
             Tip: ko.computed(() => tips[currentTipIndex() % tips.length]),
             NextTip: cycleTipIndex.bind(1),
-            PreviousTip: cycleTipIndex.bind(-1)
+            PreviousTip: cycleTipIndex.bind(-1),
+            SaveAndClose: params.encounterCommander.HideSettings
         }
     }
 }
